cleaned_requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
beartype==0.16.4
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
distro==1.8.0
einops==0.7.0
evdev==1.6.1
exceptiongroup==1.2.0
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
gTTS==2.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
inflect==7.0.0
Jinja2==3.1.2
joblib==1.3.2
keyboard==0.13.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
openai==0.28.0
openai-whisper==20231117
packaging==23.2
Pillow==10.1.0
platformdirs==4.0.0
playsound==1.3.0
pocketsphinx==5.0.2
pooch==1.8.0
progressbar==2.5
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycairo==1.25.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyGetWindow==0.0.9
PyGObject==3.46.0
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-xlib==0.33
python3-xlib==0.15
pytweening==1.0.7
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rotary-embedding-torch==0.4.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
semantic-version==2.10.0
setuptools-rust==1.8.1
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soxr==0.3.7
SpeechRecognition==3.10.0
sympy==1.12
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.1.1
tortoise-tts==3.0.0
tqdm==4.66.1
transformers==4.31.0
triton==2.1.0
typing_extensions==4.8.0
Unidecode==1.3.7
urllib3==2.1.0
yarl==1.9.3
